Add some more cases where parentheses are needed in text output.
Fractions need parentheses around the numerator or denominator if
anything is after a factorial or a SupSub.  For example, 2/(3!4!) or
2/(2^2 3^2).

SupSubs need parentheses on the superscripts or subscripts if anything
is after a factorial. For example, 2^(2!3!)

Also, a live fraction should be allowed to be typed after a factorial.
For example, if 2! is typed and then the forward slash is typed it
should result in the fraction with 2! in the numerator and the cursor in
the denominator.  Currenlty the 2! is in front of the fraction and the
cursor is in the numerator.

This fixes https://github.com/openwebwork/webwork2/issues/2821.